saving homepage + js before merge
