+++ Develop/server.js
@@ -12,8 +12,8 @@ const app = express();
 const PORT = process.env.PORT || 3001;
 
 // adds the functionality of helpers to handlebars engine
-const helpers = require('./utils/helpers');
-const hbs = exphbs.create({ helpers });
+// const helpers = require('./public/js/helpers');
+const hbs = exphbs.create({ });//helpers 
 
 // constructing session/cookie
 const sess = {
