+++ Develop/views/homepage.handlebars
@@ -5,31 +5,31 @@
 
 <content class="flex flex-wrap h-full">
 
-    <aside class="min-w-max ml-8">       
-        <div class="px-8 flex h-full flex-col bg-gray-300 items-center space-y-14">
+    <aside class="min-w-max ml-8 flex-grow">       
+        <div class="px-8 flex h-full flex-col bg-gray-100 bg-opacity-50 border-2 border-black items-center space-y-14">
              
-            <div id="user-pic" class="w-40 border-8 mt-8">
-            <img src="icon.png" alt="Profile Picture">
+            <div id="user-pic" class="w-40 border-4 border-black mt-8">
+            <img src="/imgs/icon.png" alt="Profile Picture">
             </div>
 
             <div id="items" class="text-4xl space-y-24">
 
                 {{!-- displays username in side bar --}}
                 <div>
-                    <p>{{name}}</p>
+                    <p>{{user.name}}</p>
                 </div>
                   
                 {{!-- Sidebar buttons --}}
                 <div id="new-entry">
-                    <button class="p-2 border border-black"><i class='fa fa-plus mr-2'></i>New Entry</button>
+                    <button class="p-2 border border-black hover:bg-gray-50"><i class='fa fa-plus mr-2'></i>New Entry</button>
                 </div>
 
                 <div id="past-entries">
-                    <button class="p-2 border border-black"><i class='fa fa-book mr-2'></i>Past Entries</button>
+                    <button class="p-2 border border-black hover:bg-gray-50"><i class='fa fa-book mr-2'></i>Past Entries</button>
                 </div>
 
                 <div id="user-settings">
-                    <button class="p-2 border border-black mx-8"><i class='fa fa-gear mr-2'></i>User Settings</button>
+                    <button class="p-2 border border-black mx-8 hover:bg-gray-50"><i class='fa fa-gear mr-2'></i>User Settings</button>
                 </div>
             </div>
         </div>
@@ -37,18 +37,18 @@
 
     {{!-- empty div created to place the pie chart in --}}
     <section id="chart-container" class="flex-grow flex flex-col items-center justify-center">
-        <div id="chart"></div>
+        <div id="chart" onload="chartGenerate({{journal}})"></div>
     </section>
 
     {{!-- box to place the user's last journal entry in --}}
     <section id="box-container" class="flex h-full flex-grow flex-col space-y-20 mr-8 items-center">
-        <div class="border border-black h-56 w-56 space-y-12">
+        <div class="border border-black h-56 w-56 space-y-12 bg-gray-100 bg-opacity-50">
             <h1 class="text-2xl mt-4">Your last journal entry</h1>
-            <p class="text-xl">{{journal.[journal.length].topic}}</p>
+            <p class="text-xl">{{user.journal.[journal.length].topic}}</p>
         </div>
 
         {{!-- box created to hold the generated mood calendar --}}
-        <div class="h-56 w-56 space-y-2 border-r-2 border-l-2 border-t-2 border-black">
+        <div class="h-56 w-56 space-y-2 border-r-2 border-l-2 border-t-2 border-black bg-gray-100 bg-opacity-50">
             <h2 class="mt-2">Monthly Mood Calendar</h2>
 
             {{!-- calendar labels --}}
@@ -67,10 +67,10 @@
         {{!-- badge box to present the user their most recent badge and also click to
               see their total badges (not implemented) --}}
         <button id="badge-button">
-        <div class="flex flex-col border border-black h-56 w-56 items-center">
+        <div class="flex flex-col border border-black h-56 w-56 items-center hover:opacity-80 bg-gray-100 bg-opacity-50">
             <h3 class="text-2xl mt-4">Badges</h3>
             <p class="mt-2">Your most recent badge is..</p>
-            <img class="h-24 w-24" src="badge-example.png" alt="Recent Example Badge">
+            <img class="h-24 w-24" src="/imgs/badge-example.png" alt="Recent Example Badge">
             <p>for logging 7 days of consecutive entries</p>
         </div>
     </button>
@@ -78,6 +78,7 @@
 
 </content>
 
+<script src="./js/helpers.js"></script>
 {{!-- calls the helper functions to generate the pie chart and mood calendar --}}
-{{{gen_chart user}}}
-{{{gen_calendar journal}}}
+{{!-- {{{ user.journal}}}
+{{{gen_calendar user.journal}}}  --}}
